refactor: extract dropdown menu to separate client component
